obs-bridge/requirements.txt
websockets==13.1
obs-websocket-py==1.0
